PF/PB/requirements.txt
django
virtualenv
djangorestframework_simplejwt
Pillow
python-dateutil
coreapi
djangorestframework
django-cors-headers